+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 ## WIP
 
-- Document formatting is now available for `.tra`, `.msg`, and `.2da` files.
+- Document formatting is now available for `.tra`, `.msg`, `.2da`, and `scripts.lst` files.
 - New: 2DA tables have improved highlighting.
 
 ## 3.6.0